requirements.txt
acme==4.1.1
asgiref==3.8.1
boto3==1.39.9
botocore==1.39.9
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
ConfigArgParse==1.7.1
configobj==5.0.9
cryptography==45.0.5
distro==1.9.0
Django==5.1.5
django-countries==7.6.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
gunicorn==22.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
josepy==2.1.0
packaging==24.2
parsedatetime==2.6
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.1
sentry-sdk==2.35.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
